move anaconda_project.plugins to anaconda_project.requirements_registry
