+++ anaconda_project/requirements_registry/providers/test/test_redis_provider.py
@@ -17,10 +17,10 @@ from anaconda_project.internal.test.tmpfile_utils import (with_directory_content
 from anaconda_project.test.environ_utils import minimal_environ, strip_environ
 from anaconda_project.local_state_file import DEFAULT_LOCAL_STATE_FILENAME
 from anaconda_project.local_state_file import LocalStateFile
-from anaconda_project.plugins.registry import RequirementsRegistry
-from anaconda_project.plugins.requirement import UserConfigOverrides
-from anaconda_project.plugins.providers.redis import RedisProvider
-from anaconda_project.plugins.requirements.redis import RedisRequirement
+from anaconda_project.requirements_registry.registry import RequirementsRegistry
+from anaconda_project.requirements_registry.requirement import UserConfigOverrides
+from anaconda_project.requirements_registry.providers.redis import RedisProvider
+from anaconda_project.requirements_registry.requirements.redis import RedisRequirement
 from anaconda_project.prepare import prepare_without_interaction, unprepare
 from anaconda_project import provide
 from anaconda_project.project_file import DEFAULT_PROJECT_FILENAME
@@ -168,7 +168,8 @@ def _monkeypatch_can_connect_to_socket_to_succeed(monkeypatch):
         can_connect_args['timeout_seconds'] = timeout_seconds
         return True
 
-    monkeypatch.setattr("anaconda_project.plugins.network_util.can_connect_to_socket", mock_can_connect_to_socket)
+    monkeypatch.setattr("anaconda_project.requirements_registry.network_util.can_connect_to_socket",
+                        mock_can_connect_to_socket)
 
     return can_connect_args
 
@@ -205,7 +206,8 @@ def _monkeypatch_can_connect_to_socket_on_nonstandard_port_only(monkeypatch, rea
         else:
             return real_can_connect_to_socket(host, port, timeout_seconds)
 
-    monkeypatch.setattr("anaconda_project.plugins.network_util.can_connect_to_socket", mock_can_connect_to_socket)
+    monkeypatch.setattr("anaconda_project.requirements_registry.network_util.can_connect_to_socket",
+                        mock_can_connect_to_socket)
 
     return can_connect_args_list
 
@@ -217,7 +219,7 @@ def test_prepare_and_unprepare_local_redis_server(monkeypatch):
         print("Cannot start redis-server on Windows")
         return
 
-    from anaconda_project.plugins.network_util import can_connect_to_socket as real_can_connect_to_socket
+    from anaconda_project.requirements_registry.network_util import can_connect_to_socket as real_can_connect_to_socket
 
     can_connect_args_list = _monkeypatch_can_connect_to_socket_on_nonstandard_port_only(monkeypatch,
                                                                                         real_can_connect_to_socket)
@@ -273,7 +275,7 @@ def test_prepare_and_unprepare_local_redis_server_with_failed_unprovide(monkeypa
         print("Cannot start redis-server on Windows")
         return
 
-    from anaconda_project.plugins.network_util import can_connect_to_socket as real_can_connect_to_socket
+    from anaconda_project.requirements_registry.network_util import can_connect_to_socket as real_can_connect_to_socket
 
     _monkeypatch_can_connect_to_socket_on_nonstandard_port_only(monkeypatch, real_can_connect_to_socket)
 
@@ -304,7 +306,7 @@ def test_prepare_and_unprepare_two_local_redis_servers_with_failed_unprovide(mon
         print("Cannot start redis-server on Windows")
         return
 
-    from anaconda_project.plugins.network_util import can_connect_to_socket as real_can_connect_to_socket
+    from anaconda_project.requirements_registry.network_util import can_connect_to_socket as real_can_connect_to_socket
 
     _monkeypatch_can_connect_to_socket_on_nonstandard_port_only(monkeypatch, real_can_connect_to_socket)
 
@@ -337,7 +339,7 @@ def test_prepare_local_redis_server_twice_reuses(monkeypatch):
         print("Cannot start redis-server on Windows")
         return
 
-    from anaconda_project.plugins.network_util import can_connect_to_socket as real_can_connect_to_socket
+    from anaconda_project.requirements_registry.network_util import can_connect_to_socket as real_can_connect_to_socket
 
     can_connect_args_list = _monkeypatch_can_connect_to_socket_on_nonstandard_port_only(monkeypatch,
                                                                                         real_can_connect_to_socket)
@@ -409,7 +411,7 @@ def test_prepare_local_redis_server_times_out(monkeypatch, capsys):
         print("Cannot start redis-server on Windows")
         return
 
-    from anaconda_project.plugins.network_util import can_connect_to_socket as real_can_connect_to_socket
+    from anaconda_project.requirements_registry.network_util import can_connect_to_socket as real_can_connect_to_socket
 
     _monkeypatch_can_connect_to_socket_on_nonstandard_port_only(monkeypatch, real_can_connect_to_socket)
 
@@ -476,7 +478,8 @@ def _monkeypatch_can_connect_to_socket_always_succeeds_on_nonstandard(monkeypatc
         can_connect_args_list.append(can_connect_args)
         return port != 6379
 
-    monkeypatch.setattr("anaconda_project.plugins.network_util.can_connect_to_socket", mock_can_connect_to_socket)
+    monkeypatch.setattr("anaconda_project.requirements_registry.network_util.can_connect_to_socket",
+                        mock_can_connect_to_socket)
 
     return can_connect_args_list
 
@@ -551,7 +554,8 @@ def _monkeypatch_can_connect_to_socket_always_fails(monkeypatch):
     def mock_can_connect_to_socket(host, port, timeout_seconds=0.5):
         return False
 
-    monkeypatch.setattr("anaconda_project.plugins.network_util.can_connect_to_socket", mock_can_connect_to_socket)
+    monkeypatch.setattr("anaconda_project.requirements_registry.network_util.can_connect_to_socket",
+                        mock_can_connect_to_socket)
 
 
 def test_fail_to_prepare_local_redis_server_scope_system(monkeypatch, capsys):
@@ -611,7 +615,7 @@ service_options:
 def _fail_to_prepare_local_redis_server_exec_fails(monkeypatch, capsys, logfile_fail_mode):
     # this test will fail if you don't have Redis installed, since
     # it actually starts it.
-    from anaconda_project.plugins.network_util import can_connect_to_socket as real_can_connect_to_socket
+    from anaconda_project.requirements_registry.network_util import can_connect_to_socket as real_can_connect_to_socket
 
     _monkeypatch_can_connect_to_socket_on_nonstandard_port_only(monkeypatch, real_can_connect_to_socket)
 
@@ -694,7 +698,7 @@ def test_fail_to_prepare_local_redis_server_exec_fails_logfile_is_dir(monkeypatc
 
 
 def test_fail_to_prepare_local_redis_server_not_on_path(monkeypatch, capsys):
-    from anaconda_project.plugins.network_util import can_connect_to_socket as real_can_connect_to_socket
+    from anaconda_project.requirements_registry.network_util import can_connect_to_socket as real_can_connect_to_socket
 
     _monkeypatch_can_connect_to_socket_on_nonstandard_port_only(monkeypatch, real_can_connect_to_socket)
