+++ setup.py
@@ -559,6 +559,7 @@ setup(name='anaconda-project',
           'bin/anaconda-project'
       ],
       packages=[
-          'anaconda_project', 'anaconda_project.internal', 'anaconda_project.internal.cli', 'anaconda_project.plugins',
-          'anaconda_project.plugins.providers', 'anaconda_project.plugins.requirements'
+          'anaconda_project', 'anaconda_project.internal', 'anaconda_project.internal.cli',
+          'anaconda_project.requirements_registry', 'anaconda_project.requirements_registry.providers',
+          'anaconda_project.requirements_registry.requirements'
       ])
